fix(aero-machine): correct vbe_lfb_base test types
Machine::vbe_lfb_base() returns a u32. Recent test additions compared it against u64 values, causing aero-machine tests to fail to compile.\n\nNormalize the assertions to compare u32 values.

+++ crates/aero-machine/src/lib.rs
@@ -11752,7 +11752,7 @@ mod tests {
         .unwrap();
         assert_eq!(
             headless.vbe_lfb_base(),
-            u64::from(firmware::video::vbe::VbeDevice::LFB_BASE_DEFAULT)
+            firmware::video::vbe::VbeDevice::LFB_BASE_DEFAULT
         );
 
         // When VGA is enabled, the BIOS should report the legacy MMIO-mapped SVGA base.
@@ -11767,7 +11767,7 @@ mod tests {
             ..Default::default()
         })
         .unwrap();
-        assert_eq!(vga.vbe_lfb_base(), u64::from(aero_gpu_vga::SVGA_LFB_BASE));
+        assert_eq!(vga.vbe_lfb_base(), aero_gpu_vga::SVGA_LFB_BASE);
     }
 
     #[test]
@@ -11791,9 +11791,8 @@ mod tests {
             .unwrap_or(0);
         assert_ne!(bar1_base, 0, "AeroGPU BAR1 should be assigned by BIOS POST");
 
-        let expected = u64::from(
-            u32::try_from(bar1_base + VBE_LFB_OFFSET as u64).expect("LFB base should fit in u32"),
-        );
+        let expected = u32::try_from(bar1_base + VBE_LFB_OFFSET as u64)
+            .expect("LFB base should fit in u32");
         assert_eq!(m.vbe_lfb_base(), expected);
     }
